nemesis: fixed mismatch in recv_seqno and send_seqno types

+++ src/mpid/ch3/channels/nemesis/include/mpidi_ch3_pre.h
@@ -22,7 +22,7 @@ typedef struct MPIDI_CH3I_VC
     struct MPID_Request *recv_active;
 
     int is_local;
-    int send_seqno;
+    unsigned short send_seqno;
     MPID_nem_fbox_mpich2_t *fbox_out;
     MPID_nem_fbox_mpich2_t *fbox_in;
     MPID_nem_queue_ptr_t recv_queue;
